+++ doc/release_notes.md
@@ -10,6 +10,10 @@ Please list contributions, add reference to PRs if present.
 
 * Implement **deliverability** constraint to allow use of electricity for hydrogen production only if produced within the same grid region. [PR #122](https://github.com/open-energy-transition/efuels-supply-potentials/pull/122)
 
+* Integrate **gas-input** of steel and cement carbon capture technologies. [PR #121](https://github.com/open-energy-transition/efuels-supply-potentials/pull/121)
+
+* Add **pinned environment files** for different OS for reproducibility [PR #118](https://github.com/open-energy-transition/efuels-supply-potentials/pull/118)
+
 * Restore tax credits for batteries and correct application for DAC. [PR #120](https://github.com/open-energy-transition/efuels-supply-potentials/pull/120)
 
 * Fix eligibility criteria for CC retrofits to comply with CO2 parallel routes. [PR #116](https://github.com/open-energy-transition/efuels-supply-potentials/pull/116)